Criando o modelo para as repostas.
close #48
