+++ quiz/base/admin/__init__.py
@@ -1,3 +1,3 @@
-from quiz.base.admin.pergunta import PeguntaAdmin
+from quiz.base.admin.pergunta_e_resposta import PeguntaAdmin, RepostasAdmin
 from quiz.base.admin.user import UserAdmin
 from quiz.base.admin.aluno import AlunoAdmin
