+++ quiz/base/admin/aluno.py
@@ -5,4 +5,4 @@ from quiz.base.models import Aluno
 
 @admin.register(Aluno)
 class AlunoAdmin(admin.ModelAdmin):
-    list_display = ('id', 'nome', 'email')
+    list_display = ('id', 'nome', 'email', 'criado_em')
