+++ quiz/base/admin/pergunta_e_resposta.py
@@ -1,0 +1,13 @@
+from django.contrib import admin
+
+from quiz.base.models import Pergunta, Resposta
+
+
+@admin.register(Pergunta)
+class PeguntaAdmin(admin.ModelAdmin):
+    list_display = ('id', 'enunciado', 'disponivel')
+
+
+@admin.register(Resposta)
+class RepostasAdmin(admin.ModelAdmin):
+    list_display = ('id', 'aluno', 'pergunta', 'pontos', 'respondida_em')
