+++ quiz/base/models/__init__.py
@@ -1,3 +1,3 @@
 from quiz.base.models.user import User, UserManager
-from quiz.base.models.pergunta import Pergunta
 from quiz.base.models.aluno import Aluno
+from quiz.base.models.pergunta_e_resposta import Pergunta, Resposta
