--- quiz/base/models/pergunta.py
@@ -1,16 +1,0 @@
-from django.db import models
-
-
-class Pergunta(models.Model):
-    enunciado = models.TextField()
-    disponivel = models.BooleanField(default=False)
-    alternativas = models.JSONField()
-    alternativas_correta = models.IntegerField(choices=[
-        (0, 'A'),
-        (1, 'B'),
-        (2, 'C'),
-        (3, 'D'),
-    ])
-
-    def __str__(self):
-        return self.enunciado
